created controller, entity, repository, service

+++ src/main/java/io/zipcoder/crudapp/repository/PersonRepository.java
@@ -3,6 +3,5 @@ package io.zipcoder.crudapp.repository;
 import io.zipcoder.crudapp.entity.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person, Long> {
-
+public interface PersonRepository extends CrudRepository<Person, Integer> {
 }
